+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.7.0] - 2023-09-18
+
+## Added
+- `sendOrderWithoutSecretKey` method to `MultisigWallet`
+
+## Fixed
+- Uri encode get method name for `TonClient4`
+
+
 ## [13.6.1] - 2023-08-24
 
 ## Fixed